Split up init.lua Not tested yet. I've simply transferred most of the code to 6 other files.
